+++ src/test/java/de/alpharogroup/reflection/ReflectionExtensionsTest.java
@@ -501,6 +501,7 @@ public class ReflectionExtensionsTest
 	{
 		Person expected;
 		Person actual;
+
 		final Class<Person> clazz = Person.class;
 		actual = ReflectionExtensions.newInstance(clazz);
 		assertNotNull(actual);
@@ -533,6 +534,22 @@ public class ReflectionExtensionsTest
 		assertEquals(expected, actual);
 		expected = A.builder().build();
 		assertEquals(expected, actual);
+		// new scenario with object array...
+		Integer[] integerArray = ArrayFactory.newArray(1, 2, 3);
+		actual = ReflectionExtensions.newInstance(integerArray);
+		assertNotNull(actual);
+		assertTrue(actual instanceof Integer[]);
+		Integer[] integerArrayActual = (Integer[])actual;
+		Integer[] integerArrayExpected = ArrayFactory.newArray(null, null, null);
+		assertTrue(Arrays.deepEquals(integerArrayActual, integerArrayExpected));
+		// new scenario with primitive array...
+		int[] intArray = ArrayFactory.newIntArray(1, 2, 3);
+		actual = ReflectionExtensions.newInstance(intArray);
+		assertNotNull(actual);
+		assertTrue(actual instanceof int[]);
+		int[] intArrayActual = (int[])actual;
+		int[] intArrayExpected = ArrayFactory.newIntArray(0, 0, 0);
+		assertTrue(Arrays.equals(intArrayActual, intArrayExpected));
 	}
 
 	/**
@@ -552,30 +569,6 @@ public class ReflectionExtensionsTest
 		expected.setMarried(null);
 		expected.setName(null);
 		expected.setNickname(null);
-		assertEquals(expected, actual);
-	}
-
-	/**
-	 * Test method for {@link ReflectionExtensions#setFieldValue(Object, String, Object)}.
-	 *
-	 * @throws NoSuchFieldException
-	 *             is thrown if no such field exists.
-	 * @throws SecurityException
-	 *             is thrown if a security manager says no.
-	 * @throws IllegalAccessException
-	 *             is thrown if an illegal on create an instance or access a method.
-	 */
-	@SuppressWarnings("deprecation")
-	@Test
-	public void testSetFieldValueObject()
-		throws NoSuchFieldException, SecurityException, IllegalAccessException
-	{
-		String expected;
-		String actual;
-		final Person person = Person.builder().name("Alex").build();
-		expected = "Leo";
-		ReflectionExtensions.setFieldValue(person, "name", "Leo");
-		actual = person.getName();
 		assertEquals(expected, actual);
 	}